Migrate build to ESP-IDF 6.0.1
Port the firmware to ESP-IDF v6.0.1 (from 5.3) / GCC 15. Validated with a
green native idf.py build (no PlatformIO platform ships IDF 6.0 yet).

Component-manager extractions (removed from IDF core in 6.0):
- esp-mqtt -> managed dep espressif/mqtt (new code/main/idf_component.yml)
- cJSON    -> managed dep espressif/cjson; REQUIRES json -> cjson in the
  MQTT and webhook components

Driver-component split (6.0 enforces include<->requires): add explicit
esp_driver_* (gpio/rmt/spi/ledc/sdmmc/sdspi), mdns, esp_wifi, nvs_flash, ...
to the affected component CMakeLists and to main.

API / source updates for 6.0 + GCC 15:
- esp_vfs_fat_register() now takes a config struct -> esp_vfs_fat_register_cfg()
- private sdmmc_common.h -> public sdmmc_cmd.h
- HSPI_HOST -> SPI3_HOST; SOC_RMT_* caps dropped -> ESP32 fallbacks;
  rmt_tx_channel_config_t::intr_priority added
- WIFI_REASON_NOT_AUTHED -> WIFI_REASON_ASSOC_NOT_AUTHED (version-gated)
- GCC-15: cross-enum comparison cast, transposed calloc args
- native-build board define + -Wno-error injected via CMake (non-PlatformIO)

Submodule bumps (latest stable, carry the 6.0/GCC-15 fixes):
- esp32-camera v2.0.6 -> v2.1.6
- esp-tflite-micro v1.3.1 -> v1.3.5 (fixes std::is_pod under -std=gnu++26)
- esp-nn v1.1.0 -> v1.2.0
- esp-protocols/mdns v1.4.3 -> v1.11.1

+++ code/components/jomjol_controlGPIO/server_GPIO.cpp
@@ -108,7 +108,10 @@ void GpioPin::init()
     gpio_config(&io_conf);
 
 //    if (_interruptType != GPIO_INTR_DISABLE) {                // ohne GPIO_PIN_MODE_EXTERNAL_FLASH_WS281X, wenn das genutzt wird, dann soll auch der Handler hier nicht initialisiert werden, da das dann über SmartLED erfolgt.
-    if ((_interruptType != GPIO_INTR_DISABLE) && (_interruptType != GPIO_PIN_MODE_EXTERNAL_FLASH_WS281X)) {
+    // Note: GPIO_PIN_MODE_EXTERNAL_FLASH_WS281X is a gpio_pin_mode_t value compared numerically
+    // against the gpio_int_type_t interrupt type; cast to int to keep the original behavior
+    // (GCC 15 rejects the cross-enum comparison under -Werror=enum-compare).
+    if ((_interruptType != GPIO_INTR_DISABLE) && ((int)_interruptType != (int)GPIO_PIN_MODE_EXTERNAL_FLASH_WS281X)) {
         //hook isr handler for specific gpio pin
         ESP_LOGD(TAG, "GpioPin::init add isr handler for GPIO %d", _gpio);
         gpio_isr_handler_add(_gpio, gpio_isr_handler, (void*)&_gpio);
